+++ packages/antd/src/templates/FieldTemplate/index.js
@@ -66,13 +66,15 @@ const FieldTemplate = ({
           colon={colon}
           extra={description}
           hasFeedback={schema.type !== "array" && schema.type !== "object"}
-          help={(!!rawHelp && help) || (!!rawErrors && renderFieldErrors())}
+          help={
+            (!!rawHelp && help) || (!!rawErrors?.length && renderFieldErrors())
+          }
           htmlFor={id}
           label={displayLabel && label}
           labelCol={labelCol}
           required={required}
           style={wrapperStyle}
-          validateStatus={rawErrors ? "error" : undefined}
+          validateStatus={rawErrors?.length ? "error" : undefined}
           wrapperCol={wrapperCol}
         >
           {children}